feat(stonedb): formate code with clang-format (#7)

+++ storage/stonedb/compress/code_stream.h
@@ -48,7 +48,7 @@ class CoderStream : protected ArithCoder {
   void InitCompress() { ArithCoder::InitCompress(); }
   void Encode(BaseT low, BaseT high, BaseT total) {
     CprsErr err = ArithCoder::ScaleRange(str, low, high, total);
-    if (err) throw err;
+    if (static_cast<int>(err)) throw err;
   }
   void EndCompress() { ArithCoder::EndCompress(str); }
   // decompression methods
@@ -56,32 +56,32 @@ class CoderStream : protected ArithCoder {
   BaseT GetCount(BaseT total) { return ArithCoder::GetCount(total); }
   void Decode(BaseT low, BaseT high, BaseT total) {
     CprsErr err = ArithCoder::RemoveSymbol(str, low, high, total);
-    if (err) throw err;
+    if (static_cast<int>(err)) throw err;
   }
 
   // uniform compression and decompression
   template <class T>
   void EncodeUniform(T val, T maxval, uint bitmax) {
     CprsErr err = ArithCoder::EncodeUniform<T>(str, val, maxval, bitmax);
-    if (err) throw err;
+    if (static_cast<int>(err)) throw err;
   }
 
   template <class T>
   void EncodeUniform(T val, T maxval) {
     CprsErr err = ArithCoder::EncodeUniform<T>(str, val, maxval);
-    if (err) throw err;
+    if (static_cast<int>(err)) throw err;
   }
 
   template <class T>
   void DecodeUniform(T &val, T maxval, uint bitmax) {
     CprsErr err = ArithCoder::DecodeUniform<T>(str, val, maxval, bitmax);
-    if (err) throw err;
+    if (static_cast<int>(err)) throw err;
   }
 
   template <class T>
   void DecodeUniform(T &val, T maxval) {
     CprsErr err = ArithCoder::DecodeUniform<T>(str, val, maxval);
-    if (err) throw err;
+    if (static_cast<int>(err)) throw err;
   }
 };
